mailbox: Add Renesas rcar driver
Renesas Gen 3 platforms have a Multifunctionnal Interface (MFIS)
which provides an interface between the AP System and the Arm Realtime
system. This interface supports issuing of interrupts for the
Arm Realtime Core domain by the AP System Core domain and of interrupts
for the AP System Core domain by the Arm Realtime Core domain.
There is up to 8 interrupts in each direction available.
This driver only use 2 interrupts in each direction. Tx and Tx ack,
Rx and Rx ack.

Each interrupt is on a 32bit registers where bit(0) is used to trigger the
interrupt, and the bit 15 to 1 can be used as general purpose.
The bit 1 is used here to mark a channel busy while transmitting data,
and waiting for a Tx ack where we clear this busy flag.